From Didier Gautheron via bug 3391:
when loading files > 50 MB wireshark redraws the first pane on each
update_progress_dlg(). If auto_scroll_live is not set that's mean it redraws
the same rows again and again.

The patch attached only redraws it once or if cf->displayed_count < 1000, in
case you have a very big screen.


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@28475 f5534014-38df-0310-8fa8-9805f1628bb7