From Adam Nielsen via bug 3659:
Here is a patch against SVN that implements preliminary support for USB HID
devices.  At the moment it only dissects the initial set up packets (which you
will see if you hotplug a keyboard or mouse.)

The patch also fixes a minor bug in the USB dissector code which reported
certain packets as malformed due to miscounting bytes, and it reports control
packet data and padding data in a more user-friendly manner.


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@29256 f5534014-38df-0310-8fa8-9805f1628bb7