Move the definition of the uuid, and registration of its name, for the standard ndr transfer syntax from the epm dissector to packet-dcerpc.c
Add a new transfer syntax : ndr64. This is a new syntax with different
scalar sizes and different alignment rules compared to normal ndr.
It is negotiated and used between w2k8 and samba4 boxens and one may 
assume, future versions of windows as well.

We need to associate the transfer syntax with the bind information since
the transfer syntax will change the packet encoding rules for the 
protocol.
For example, SAMR, as well as all other interfaces support both syntaxes 
and are thus encoded differently, wiht different alignments depending on 
which transfer was negotioated during the bind.



This will require additional changes to the dcerpc helpers and also to 
pidl.




git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@30209 f5534014-38df-0310-8fa8-9805f1628bb7