From Sven Eckelmann via but 2631:
The wireless meshing protocol B.A.T.M.A.N. Advanced changed their packet format
in such a way that now versions can be identified and so correct dissection of
the packets can be supported by wireshark.
Since it is a ever moving target it is very possible that the packet format is
changing slightly. The dissector was written in such a way that new version can
be supported relative easy.
I hope that it sufficient for the inclusion in wireshark.

I tried to fuzzing it some hours and no error was reported.


From me:

Initialize our dissector handles.

Merge packet-batadv.h into packet-batadv.c. It isn't included anywhere else.

Fuzz 500 passes using attached capture files.


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@33052 f5534014-38df-0310-8fa8-9805f1628bb7