From Pascal Quantin:
The decoder expects to do not have any extra byte after the RAR bodies.

But according to 3GPP 36.321 8.8.0 chapter 6.1.5, the eNodeB is allowed to put padding at the end of the RAR PDU:

" A MAC PDU consists of a MAC header and zero or more MAC Random Access Responses (MAC RAR) and optionally padding as described in figure 6.1.5-4.
[...]

Padding may occur after the last MAC RAR. Presence and length of padding is implicit based on TB size, size of MAC header and number of RARs."


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@33783 f5534014-38df-0310-8fa8-9805f1628bb7