Support multiple header fields with the same abbreviation in custom columns.
We have some different fields using the same abbreviation (e.g "eth.dst"
used in both eth and tte), and this patch will fetch values from all fields.
When using occurrences the entries listed first is from the field registered
last when starting Wireshark, and not ordered from the occurrence in the
packet, but I don't see how we can easily fix this.


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@34513 f5534014-38df-0310-8fa8-9805f1628bb7