From Holger Freyther via https://bugs.wireshark.org/bugzilla/show_bug.cgi?id=5320 :
Doing a Dumper:close() and then forcing the dumper to be garbage collected will
cause a segfault.

The code attempts to handle the GC that could follow but fails to do it
properly. Assigning Dumper dp = NULL will only have an affect within the scope
of this method... Assumingly this should be a *dp = NULL.


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@34597 f5534014-38df-0310-8fa8-9805f1628bb7