Tighten up code dissecting message header.
This corrects the specific issue reported in Bug #3317
wherein the dissector decided there was a valid but unknown
header when dissecting a binary (non text) message with a ":"
as the last byte.

https://bugs.wireshark.org/bugzilla/show_bug.cgi?id=3317

Note that a larger issue remains: the IMF dissector
presumably shouldn't really even try to dissect a
binary payload (which is proably encrypted text).


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@35017 f5534014-38df-0310-8fa8-9805f1628bb7