From Neil Piercy via https://bugs.wireshark.org/bugzilla/show_bug.cgi?id=3301 :
The information which is used to determine which sub-dissector to use for the
various Data messages within an SCCP connection is only present within the
initial Connection Request, so even with connection tracking on, unless the
trace contains the Connection Request no sub-dissector is called. It is common
for traces to only contain a single carried protocol anyway - e.g. RANAP.

The supplied patch adds a user preference for a "default payload"
sub-dissector, which is called in preference to the Data dissector if nothing
else has claimed the packet first.


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@35098 f5534014-38df-0310-8fa8-9805f1628bb7