Put the rate in for the MCS field, if all the information needed to calculate the rate is present.
Don't put the rate in for MCS indices in the Rate field - we need not
only the MCS index but the bandwidth (20 MHz vs. 40 MHz) and the length
of the guard interval.  Note in a comment that those other items *might*
be available from some other radiotap fields.

Expand a comment on the hack to try to figure out whether a Rate field
with the 0x80 bit set is a FreeBSD MCS index or a high data rate.


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@36916 f5534014-38df-0310-8fa8-9805f1628bb7