From Brian Cavagnolo via https://bugs.wireshark.org/bugzilla/show_bug.cgi?id=6173 Update 802.11s packet dissecting to the ratified standard (v12.0)
[PATCH 7/9] update parsing of mesh control field

During development of the 80211s standard, a "mesh header" was conceived. This mesh header has been renamed the "mesh control field". Further, the conditions under which it is expected to appear have also changed. Specifically, the mesh control field appears in multihop action frames and mesh data frames. In the former case, it appears after the action category and action code, so no special header parsing is required to parse it.

The latter case is a bit more complicated. We know the mesh control field is present if the data frame was transmitted by a mesh STA, AND the new "Mesh Control Present" bit in the QoS control field is set.  This second thing is easy enough to check. But the first thing is not. So we continue to rely on heuristics.  Specifically, we only expect the mesh control field for valid from-ds/to-ds settings, and if the mesh control field itself is valid.

Other relevant changes in this patch include:

-- rename mesh_header to mesh_control as appropriate
-- consider the mesh control field to be part of the header when accounting for payload padding.
-- parse some of the qos fields earlier so they can be used to determine if the mesh control field is present.
-- use existing mesh control parsing code instead of duplicating it.


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@38279 f5534014-38df-0310-8fa8-9805f1628bb7