Revert rev 38340 (which was itself a reversion of 34807): if we can't find the SVN version, indicate that the SVN version is unknown.  This puts back the fix for bug 1413.
Add a new version.conf option for make-version which tell is "this is a build
from a release tarball."  When that option is present do not try to use SVN
to determine the SVN version, just use whatever SVN information shipped in the
tarball.

If version.conf is present in the source tree (as it is only in the release
branches), deliver it in the source tarball but only after setting the "this
is a release tarball" option.


All of this means that that builds from release-branch tarballs will report
the SVN version of the release tarball rather than "unknown."  This addresses
the issue reported in
http://ask.wireshark.org/questions/5376/wireshark-161-title-shows-svn-rev-unknown-from-unknown

Builds from trunk (including the source tarballs) will continue to report that
the SVN version is unknown.  (Maybe that, too, should be changed?)


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@38933 f5534014-38df-0310-8fa8-9805f1628bb7