Use ENC_ values in proto_tree_add_item() calls.  Use ENC_UTF_8 for strings, and note that, for older AFP clients and servers, we might need a way to say to use some Mac encoding instead.
Use tvb_strsize() rather than tvb_get_ephemeral_stringz() to just get
the length of a null-terminated string.

Use FT_GUID for UUIDs.

The low-order bit in the MessageBitmap in the FPGetSrvrMsg reply does
not, at least according to the current AFP spec, specify whether there's
a message at all, it specifies whether it's a server message or a login
message.  The spec *does* now mention the "message is UTF-8" bit; use
it.

Fix a blurb.


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@39063 f5534014-38df-0310-8fa8-9805f1628bb7