From Neil Piercy:
The binary display and decimal values are corrupt in the case where they are
"stitched" together from words and octets: the complete words are extracted
ready for LSB padding, but when the final octet/word is added (also so
extracted), the data already present is shifted by a whole number of octets
(not allowing for the padding), and the value is then used by the binary and
decimal display functions as if it were MSB padded. This results in both a
corrupt bit pattern and wrong padding of the bit pattern in the display.

https://bugs.wireshark.org/bugzilla/show_bug.cgi?id=6473

git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@39553 f5534014-38df-0310-8fa8-9805f1628bb7