From Pascal Quantin via bug 6593:
With Subversion 1.7, the working copy metada storage as changed (see
http://subversion.apache.org/docs/release-notes/1.7.html#wc-ng for details).
As a consequence, the file .svn/entries is no more updated and under Windows
the svnversion.h is no more regenerated (unless explicitely removed).
The attached patch fixes this issue so as to also check the .svn/wc.db file
date in the makefile rule.

Note that wc.db file must be checked before entries file as an upgrade of an
existing repository from subversion 1.6 to 1.7 leave an old entries file around
(that is no more updated).

In Makefile.am file, the svnversion.h file generation seems to be
systematically forced. So I guess Linux/Mac boxes are not impacted.


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@39910 f5534014-38df-0310-8fa8-9805f1628bb7