From Mike Morrin via https://bugs.wireshark.org/bugzilla/show_bug.cgi?id=6754 :
Due to the variable remaining_bits_len getting out of sync with bit_offset (in
one case due to a mistake in the patch for bug 6375, and in another case
pre-existing).

I have shuffled the decrements of remaining_bits_len so that they always occur
next to an increment of bit_offset, so that this type of problem is easier to
spot.

From me: convert tabs to spaces to match the rest of the file.


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@40662 f5534014-38df-0310-8fa8-9805f1628bb7