Revisit https://bugs.wireshark.org/bugzilla/show_bug.cgi?id=3161 and r39501:
Setting _XOPEN_SOURCE to 600 is only allowed on Solaris 10 if the compiler is
set to C99 mode.  Conversely (and as reported in the bug), simply defining it
(but with no value) is not allowed if the compiler *is* compiling to C99.

So, don't define _XOPEN_SOURCE at all on Solaris.  Keep defining it as 600 on
other OS's as (also) requested in that bug.

Maybe there's a cleaner way to do this but all of this is a "trickery" mess
anyway...


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@41182 f5534014-38df-0310-8fa8-9805f1628bb7