Harald Welte <laforge@gnumonks.org> via https://bugs.wireshark.org/bugzilla/show_bug.cgi?id=6792
This is a new dissector for the non-standard Ericsson OM2000 protocol, as it is
used for the OML on A-bis of Ericsson RBS 2xxx BTSs.

It also includes a dissector for a shim-layer protocol that Ericsson uses for
IP-based A-bis like the RBS 2409.  As the protocol is not publicly documented,
I have invented the name "EHDLC" (Ericsson HLDC) for it.



git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@41195 f5534014-38df-0310-8fa8-9805f1628bb7