From Weston Andros Adamson via https://bugs.wireshark.org/bugzilla/show_bug.cgi?id=7010 :
While developing SP4_MACH_CRED support for the linux nfs client, I noticed that
wireshark was displaying these bitmasks incorrectly (as fattrs instead of NFS
operations).

I also noticed that BACKCHANNEL_CTL (a required value in the mask for
SP4_MACH_CRED) wasn't in names_nfsv4_operation[], so I added it and verified
that we're not missing any other ops.


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@41851 f5534014-38df-0310-8fa8-9805f1628bb7