Get rid of the button row for the capture-file comments.  Instead, turn the whole summary dialog into an editor-like dialog with an OK and Cancel buttons (OK sets the new capture file comment, Cancel doesn't).
In order to keep the dialog the same regardless of the file type (and avoid
having a Cancel and OK button when there's no text field to edit), allow
users to create or edit capture-file comments even if the file type is not
PCAPNG (they can add a comment via the add/edit comment UI anyway).

Don't include a Clear button: the user can just Ctrl-A + backspace if they
want to do that.

Don't set the comment text to "[None]" if there's no comment, just leave it
blank.

Don't allow the user to create more than 1 Summary dialog at a time.


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@42834 f5534014-38df-0310-8fa8-9805f1628bb7