Have wtap_get_savable_file_types() take an array of encapsulations and only return file types that could handle a single file with all those encapsulations - this means that
	1) if there's more then one encapsulation, the file format has
to handle per-packet encapsulation;

	2) just because a file format handles per-packet encapsulation,
that doesn't mean that it can handle the *particular* encapsulations
being handed to it.

This fixes some cases where we were claiming that a file could be saved
in a format that doesn't actually support it (e.g., ISDN files being
reported as savable in pcap-NG format - there's no LINKTYPE_ value for
ISDN including B and D channels).


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@43300 f5534014-38df-0310-8fa8-9805f1628bb7