As per Gentoo bug 423743
	https://bugs.gentoo.org/show_bug.cgi?id=423743

"The Makefile.am claims including GLIB_LIBS when linking wireshark is
unnecessary, because wireshark links to GTK_LIBS which is a superset.
It is not actually a superset: gmodule is included in GLIB_LIBS but
not in GTK_LIBS (unless accidentally on older glibs/gtks)."

so we must explicitly link with GLIB_LIBS.

Update the comment to reflect that - and to reflect that GTK+ doesn't
necessarily run atop X11 - while we're at it.

Fixes bug 7427.

#BACKPORT


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@43561 f5534014-38df-0310-8fa8-9805f1628bb7