Eliminate a lot of function-local variables by not storing the result of (several of the) get_CDR_*() functions: instead call the function directly in the proto_tree_add_text() calls.
Eliminate a lot of global variables that are only later used in (a single)
strcmp().

All the GIOP dissectors are built-in now; don't include the (#if'd out) plugin
registration code in the generated dissector.

Try to clean up the formatting of the generated code a bit.

Regenerate the IDL dissectors; as noted in r39932, this seems to involve a
lot of function re-ordering in packet-parlay.c (making the delta bigger).


These changes decrease the time to compile packet-parlay.c by, on my system,
about half.  But it still complains:

variable tracking size limit exceeded with -fvar-tracking-assignments, retrying without

and it still takes "too long."


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@43609 f5534014-38df-0310-8fa8-9805f1628bb7