Speed up line drawing by no longer creating and destroying a cairo_t object for each line - rather create one and use it for the entire list, then destroy it.
Simplify the allocation of segments, by no longer messing around with a
static that may or may not be used.  Also, there was a good chance the
last one would be leaked (if it was not part of the stream we are
interested in).

Note that both of these improvements would also apply to tcp_graph.c -
will wait a couple of days to be sure the changes are good.


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@43654 f5534014-38df-0310-8fa8-9805f1628bb7